requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
blinkpy==0.23.0b8
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.7
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.3
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
python-telegram-bot==21.3
requests==2.28.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.11
yarl==1.9.4
